requirements.txt
langchain
fastapi
langserve
sse_starlette
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
pymupdf
arxiv
wikipedia
validators==0.28.1
huggingface_hub
streamlit